# yorindo-api/openapi.yaml
openapi: 3.0.3
info:
  title: Yorindo API
  version: 1.0.0
  license:
    name: Proprietary
    url: https://yorindo.id/license
  description: |
    Yorindo event management platform API.

    All dates are ISO 8601 UTC strings (e.g., "2026-04-15T02:00:00.000Z").
    All IDs are opaque strings.
    All errors use { error: { code, message, details[] } } format exclusively.
    Rate limit: POST /api/registrations is limited to 10 requests/hour per IP.
servers:
  - url: http://localhost:3000/api
    description: Local development (via Docker Compose)
  - url: https://api.yorindo.id/api
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-based)
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
...
  schemas:
...
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          example: 247
        totalPages:
          type: integer
          example: 13

    # ─── Error ─────────────────────────────────────────────────────────────────
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: SCREAMING_SNAKE_CASE error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error description
              example: Validation failed
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: Invalid email format

    # ─── Async Job ─────────────────────────────────────────────────────────────
...
    Event:
      type: object
      required: [id, name, slug, status, eventDate, timezone, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: event-001
        name:
          type: string
          example: Seminar ERP Jakarta
        slug:
          type: string
          example: seminar-erp-jakarta
        description:
          type: string
          example: Seminar tentang implementasi ERP di industri manufaktur
        status:
          type: string
          enum: [draft, published, active, completed, cancelled, archived]
          example: published
        eventDate:
          type: string
          format: date-time
          example: "2026-04-15T02:00:00.000Z"
        timezone:
          type: string
          enum: [Asia/Jakarta, Asia/Makassar, Asia/Jayapura]
          example: Asia/Jakarta
        capacity:
          type: integer
          minimum: 1
          example: 200
        targetCriteria:
          type: object
          additionalProperties: true
          description: Audience targeting criteria (JSONB)
        surveySchema:
          description: Survey template schema (JSONB)
        venue:
          type: string
          example: Jakarta Convention Center, Hall A
          description: Physical venue name and location
        industryTags:
          type: array
          items:
            type: string
          example: [teknologi, keuangan]
          description: Industry tags used for AI audience matching
        eventType:
          type: string
          enum: [conference, workshop, networking, seminar, webinar]
          example: conference
        topicTags:
          type: array
          items:
            type: string
          example: [fintech, digital-banking]
          description: Topic tags used for AI audience matching
        createdAt:
          type: string
          format: date-time
          example: "2026-03-19T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-03-19T10:30:00.000Z"
...
    User:
      type: object
      required: [id, email, name, role, createdAt, updatedAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
          example: admin@yorindo.id
        name:
          type: string
          example: Admin Yorindo
        role:
          type: string
          enum: [admin, staff, viewer, participant]
          example: admin
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserBody:
      type: object
      required: [email, name, role, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, staff, viewer, participant]
        password:
          type: string
          minLength: 8
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, staff, viewer, participant]
    PaginatedUsers:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
    UserEventAssignment:
      type: object
      required: [id, userId, eventId, grantedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        eventId:
          type: string
        grantedAt:
          type: string
          format: date-time

    # ─── Auth ──────────────────────────────────────────────────────────────────
...
paths:
...
  /users:
    get:
      summary: List all users (admin only)
      operationId: listUsers
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated users list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      summary: Create a new user (admin only)
      operationId: createUser
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserBody'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/{id}:
    patch:
      summary: Update a user (admin only)
      operationId: updateUser
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserBody'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      summary: Delete a user (admin only)
      operationId: deleteUser
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/{id}/events:
    get:
      summary: Get events assigned to user (admin only)
      operationId: getUserEvents
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User's assigned events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      summary: Assign event access to user (admin only)
      operationId: assignUserEvent
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId]
              properties:
                eventId:
                  type: string
      responses:
        '201':
          description: Event access assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEventAssignment'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: User or event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Assignment already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/me/assigned-events:
    get:
      summary: Get events assigned to the current user
      operationId: getCurrentUserAssignedEvents
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user's assigned events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/{id}/events/{eventId}:
    delete:
      summary: Revoke user's access to event (admin only)
      operationId: revokeUserEventAccess
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Access revoked
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...